refactor(config): split partial config types
